# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API-LIVRE",
        "version": "1.0"
    },
    "paths": {
...
        "/api/livres": {
            "get": {
                "tags": [
                    "livres"
                ],
                "summary": "liste des livres",
                "description": "liste des livres",
                "operationId": "livres_list",
                "responses": {
                    "200": {
                        "description": "Affiche la liste des livres"
                    }
                }
            },
            "post": {
                "tags": [
                    "livres"
                ],
                "summary": "Création d'un livre",
                "description": "Création d'un livre",
                "operationId": "livres",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sku": {
                                        "type": "string"
                                    },
                                    "nom": {
                                        "type": "string"
                                    },
                                    "points": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "sku": "Meth473",
                                    "nom": "Go For English",
                                    "points": 400,
                                    "user_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/livres/{id}": {
            "get": {
                "tags": [
                    "livres"
                ],
                "summary": "Atteindre un dossier",
                "description": "Obtenir les données d'un livre",
                "operationId": "livres_one",
                "responses": {
                    "200": {
                        "description": "Affichage d'un livre"
                    }
                }
            },
            "put": {
                "tags": [
                    "livres"
                ],
                "summary": "Mise à jour d'un livre",
                "description": "Mise à jour d'un livre",
                "operationId": "livres_update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sku": {
                                        "type": "string"
                                    },
                                    "nom": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "points": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "sku": "Meth473",
                                    "nom": "Go  English",
                                    "points": 400,
                                    "user_id": 2
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    }
}